requirements.txt
streamlit
gspread
google-auth
plotly
anthropic
python-dotenv
pandas
